backend/pyproject.toml
[tool.poetry]
name = "react-django-sns"
version = "0.1.0"
description = ""
authors = ["bp20110"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
djangorestframework = "^3.15.0"
pillow = "^10.2.0"
django-cors-headers = "3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
